+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 * Added a timeout to the `lvmbeat set overwatcher` command to prevent it from hanging indefinitely.
 * Fixed an issue that would crash the alerts task when checking the internet connection if the request timed out.
+* Ensure that `CalibrationsOverwatcher.reset()` is called.
 
 
 ## 1.4.0 - January 1, 2025